requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.15.1
frozenlist==1.4.1
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
inflection==0.5.1
kombu==5.3.7
Markdown==3.6
multidict==6.0.5
mypy-extensions==1.0.0
numpy==2.0.0
openai==1.35.3
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
